requirements.txt
nltk==3.2.5
pandas==1.1.5
scikit-learn==0.22.2.post1
numpy==1.19.5
scikit-surprise==1.1.1
